rem vs em for layout and temporary rollback
Updated my the responsive layout to use rem units for size and margins instead or em for better consistancy.
Rollback from openlist to classic html select, due to the openlist "trick" presumably causing problems on some webbrowsers. I intend to go back to this issue later and fix it but for the time being I opted for a roll-back on that.